+++ src/type_extenstions.ts
@@ -1,0 +1,11 @@
+import { BytecodeExporterConfig, BytecodeExporterUserConfig } from './types.js';
+
+declare module 'hardhat/types/config' {
+  interface HardhatConfig {
+    bytecodeExporter: BytecodeExporterConfig;
+  }
+
+  interface HardhatUserConfig {
+    bytecodeExporter?: BytecodeExporterUserConfig;
+  }
+}
